MDEV-26713 set console codepage to what user set in --default-character-set
If someone on whatever reasons uses --default-character-set=cp850,this will
avoid incorrect display, and inserting incorrect data.

That fix is currently Windows-only, and serves people who used combination
of chcp achieve WYSIWYG effect (although, this would mostly likely used
with utf8 in the past)

Now, --default-character-set would be a replacement for that.

Fix fs_character_set() detection of current codepage.
GetACP() is always correct and system default locale is often , but not
always - it does not take into account activeCodePage in application
manifest, and the fact that not all users use system default settings.