MDEV-27208: Implement the CRC32C() function
The SQL parser defines the function crc32() that computes the CRC-32
of a string using the ISO 3309 polynomial that is being used by zlib
and many others.

InnoDB and MyRocks use a different polynomial, which was implemented
in SSE4.2 instructions that were introduced in the
Intel Nehalem microarchitecture. This is commonly called CRC-32C.

Defining a SQL built-in function CRC32C() would for example allow
the definition of a simple SQL function that would generate a logically
empty InnoDB redo log corresponding to a particular checkpoint LSN.