MDEV-27199 Require ib_logfile0 to exist unless innodb_force_recovery=6
The primary purpose of the parameter innodb_force_recovery is to allow
data to be dumped from a corrupted database.

The setting innodb_force_recovery=6 will cause the redo log to be ignored.
This was somewhat redundant, because a similar effect could be achieved
by replacing ib_logfile0 with a logically empty log at the desired
log sequence number, or by simply deleting the log files. We do not want
to encourage users to engage in such inherently dangerous activity.

Users may still start up the server without ib_logfile0 by specifying
innodb_force_recovery=6. All data will be read-only.

Related to this, we no longer want InnoDB to start up when the log file
is missing or physically empty, and the FIL_PAGE_FILE_FLUSH_LSN field
in the first page of the system tablespace will no longer be written
on shutdown. That field was used as the log sequence number for a new
ib_logfile0 when the InnoDB redo log was missing at startup.